refactor: move textual UI into ui/tui package and update entry points and module imports
